refactor: extract LevelSceneRenderer from LevelCanvas
